+++ packages/plugin-packages/mashroom-portal-remote-app-registry-k8s/src/js/jobs/ScanK8SPortalRemoteAppsBackgroundJob.ts
@@ -89,6 +89,16 @@ export default class ScanK8SPortalRemoteAppsBackgroundJob implements ScanBackgro
                                         status: 'Checking',
                                         lastCheck: Date.now(),
                                         error: null,
+                                        port,
+                                    };
+                                } else if (existingService.port !== port) {
+                                    service = {
+                                        ...existingService,
+                                        priority,
+                                        status: 'Checking',
+                                        lastCheck: Date.now(),
+                                        error: null,
+                                        port,
                                     };
                                 }
                             } else {